% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_bookmark_bad_context}
\alias{err_bookmark_bad_context}
\title{Bookmark is inside a table cell, header, footer, or text box}
\usage{
err_bookmark_bad_context(name, context)
}
\arguments{
\item{name}{Bookmark name}

\item{context}{One of: "table", "header", "footer", "textbox"}
}
\description{
Bookmark is inside a table cell, header, footer, or text box
}
