% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_bookmark_missing}
\alias{err_bookmark_missing}
\title{Bookmark not found in the Word document}
\usage{
err_bookmark_missing(name)
}
\arguments{
\item{name}{Bookmark name}
}
\description{
Bookmark not found in the Word document
}
