% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_format}
\alias{err_format}
\title{Format an error for display in the UI (message + hint on a new line if present)}
\usage{
err_format(e)
}
\arguments{
\item{e}{A condition object}
}
\description{
Format an error for display in the UI (message + hint on a new line if present)
}
