% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_hint}
\alias{err_hint}
\title{Return the hint field if the condition is a houdini_error, otherwise NULL}
\usage{
err_hint(e)
}
\arguments{
\item{e}{A condition object}
}
\description{
Return the hint field if the condition is a houdini_error, otherwise NULL
}
