% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_image_extract_failed}
\alias{err_image_extract_failed}
\title{PNG image could not be extracted from an RTF file}
\usage{
err_image_extract_failed(path, cause = NULL)
}
\arguments{
\item{path}{File path}

\item{cause}{Underlying condition or message string}
}
\description{
PNG image could not be extracted from an RTF file
}
