% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_image_inject_failed}
\alias{err_image_inject_failed}
\title{Image injection failed}
\usage{
err_image_inject_failed(bookmark, cause = NULL)
}
\arguments{
\item{bookmark}{Bookmark name}

\item{cause}{Underlying condition or message string}
}
\description{
Image injection failed
}
