% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_rtf_unreadable}
\alias{err_rtf_unreadable}
\title{RTF file cannot be read}
\usage{
err_rtf_unreadable(path, cause = NULL)
}
\arguments{
\item{path}{File path}

\item{cause}{Underlying condition or message string}
}
\description{
RTF file cannot be read
}
