% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{err_xml_inject_failed}
\alias{err_xml_inject_failed}
\title{XML table injection failed}
\usage{
err_xml_inject_failed(bookmark, cause = NULL)
}
\arguments{
\item{bookmark}{Bookmark name}

\item{cause}{Underlying condition or message string}
}
\description{
XML table injection failed
}
