% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{filter_pages}
\alias{filter_pages}
\title{Filter pages by parameter value}
\usage{
filter_pages(pages, parameters)
}
\arguments{
\item{pages}{Output of parse_rtf()}

\item{parameters}{Character vector of parameter values to include}
}
\description{
Pages with NA parameter are always kept.
}
