% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{get_table_html}
\alias{get_table_html}
\title{Generate full HTML preview for an RTF file}
\usage{
get_table_html(path, pages = NULL)
}
\arguments{
\item{path}{Path to the .rtf file}

\item{pages}{Pre-parsed RTF pages (output of parse_rtf()); parsed from path if NULL}
}
\value{
HTML string
}
\description{
Generate full HTML preview for an RTF file
}
