% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{get_table_info}
\alias{get_table_info}
\title{Get summary info about an RTF file for the UI}
\usage{
get_table_info(path)
}
\arguments{
\item{path}{Path to the .rtf file}
}
\value{
list(n_cols, n_rows, col_names, parameters, timelines)
}
\description{
Get summary info about an RTF file for the UI
}
