% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{is_image_rtf}
\alias{is_image_rtf}
\title{Check whether an RTF file contains an embedded PNG image}
\usage{
is_image_rtf(path)
}
\arguments{
\item{path}{Path to the .rtf file}
}
\value{
TRUE if a \\pngblip group is found, FALSE otherwise
}
\description{
Check whether an RTF file contains an embedded PNG image
}
