% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{slice_range}
\alias{slice_range}
\title{Resolve a start/end row range to integer indices}
\usage{
slice_range(n, row_start = NULL, row_end = NULL)
}
\arguments{
\item{n}{Number of rows available}

\item{row_start}{1-based start (NULL = 1)}

\item{row_end}{1-based end (NULL = last)}
}
\value{
Integer vector of row indices (possibly empty)
}
\description{
Resolve a start/end row range to integer indices
}
