% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{slice_rows}
\alias{slice_rows}
\title{Slice data rows by start/end range}
\usage{
slice_rows(data_rows, row_start = NULL, row_end = NULL)
}
\arguments{
\item{data_rows}{List of row objects}

\item{row_start}{1-based start (NULL = 1)}

\item{row_end}{1-based end (NULL = last)}
}
\value{
Sliced list of row objects
}
\description{
Slice data rows by start/end range
}
