% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{table_info_from_pages}
\alias{table_info_from_pages}
\title{Get summary info from already-parsed pages (used by app.R's cache)}
\usage{
table_info_from_pages(pages)
}
\arguments{
\item{pages}{Output of parse_rtf()}
}
\value{
list(n_cols, n_rows, col_names, parameters, timelines)
}
\description{
Get summary info from already-parsed pages (used by app.R's cache)
}
