% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_rtf}
\alias{parse_rtf}
\title{Parse an RTF file into a list of page objects}
\usage{
parse_rtf(path)
}
\arguments{
\item{path}{Path to the .rtf file}
}
\value{
List of page objects, each with:
\item{parameter}{character or NA}
\item{header_rows}{list of row objects}
\item{data_rows}{list of row objects}
Each row object: list(is_header, cells = list(list(text, width_twips, colspan, ...)))
}
\description{
Parse an RTF file into a list of page objects
}
